Repackaged and reformatted code
