--- AircraftDB.sublime-project
@@ -1,16 +1,0 @@
-{
-	"build_systems":
-	[
-		{
-			"name": "Test",
-			"shell_cmd": "/Users/matthewgougeon/SDR/AircraftDB-Backend/virt/bin/python Aircraft_API.py"
-		}
-	],
-	"folders":
-	[
-		{
-			"path": "."
-		}
-	],
-	"virtualenv": "/Users/matthewgougeon/SDR/AircraftDB-Backend/virt"
-}
