+++ backend/scripts/database_creation.sql
