+++ backend/scripts/flight_insert_message.sql
