+++ backend/scripts/flight_new_message.sql
