+++ backend/scripts/flight_update_aircraft.sql
