+++ backend/scripts/structure.md
